services/livekit_bridge/requirements.txt
websockets>=12.0
livekit>=0.10.0
livekit-api>=0.10.0
Pillow>=10.0.0
numpy>=1.24.0
loguru>=0.7.0

